Avoid using index name to build base index mapping (#235)
CCR depends on the index name to establish the src and dest indexes
mappings. However, when renaming a table, the base index will also
be renamed. Consequently, the base index name might be the renamed
one, making it impossible to establish the src and dest indexes
mappings. Fortunately, Doris doesn't support renaming other indexes,
so the problem only occurs with the base index.

This PR directly maps the base index to the dest base index instead
of making a comparison through the index name. Therefore, whether
the index name has been renamed or not doesn't affect the upsert
synchronization.